update the content of group manage and contract manage
