+++ src/app/business/contract-manage/contract-manage.component.html
@@ -1,1 +1,63 @@
-合同管理模块
+<div class="contract-manage-header">
+  <div class="function-module-name">
+    <span class="module-name">合同管理</span>
+  </div>
+  <div class="contract-name">中南世纪城二期</div>
+  <div class="function-list">
+    <button nz-button nzType="primary">新增</button>
+    <button nz-button nzType="primary">导入</button>
+    <button nz-button nzType="default">删除</button>
+  </div>
+  <div style="clear: both"></div>
+</div>
+<div class="left-content">
+  <nz-tree [(ngModel)]="nodes" [nzShowLine]="true" [nzDefaultExpandedKeys]="expandKeys"
+    (nzExpandChange)="mouseAction('expand',$event)" (nzDblClick)="mouseAction('dblclick',$event)"
+    (nzContextMenu)="mouseAction('contextmenu', $event)" (nzClick)="mouseAction('click',$event)">
+  </nz-tree>
+</div>
+<div class="right-content">
+  <div class="search-result">共找到
+    <strong>{{contractTotal}}</strong>个合同</div>
+  <div class="search-content">
+    <nz-table class="na-table" #rowSelectionTable [nzData]="data" (nzCurrentPageDataChange)="currentPageDataChange($event)"
+      (nzPageIndexChange)="refreshStatus()" (nzPageSizeChange)="refreshStatus()">
+      <thead>
+        <tr>
+          <th nzShowCheckbox [(nzChecked)]="allChecked" [nzIndeterminate]="indeterminate" (nzCheckedChange)="checkAll($event)"></th>
+          <th></th>
+          <th>合同名称</th>
+          <th>幢号</th>
+          <th>室号</th>
+          <th>坐落</th>
+          <th>实测土地面积</th>
+          <th>实测建筑面积</th>
+          <th>实测后总价</th>
+          <th>是否贷款</th>
+          <th>贷款银行</th>
+          <th>代办费</th>
+          <th>是否含车位</th>
+          <th>当前办理进度</th>
+        </tr>
+      </thead>
+      <tbody>
+        <tr *ngFor="let data of rowSelectionTable.data">
+          <td nzShowCheckbox [(nzChecked)]="data.checked" [nzDisabled]="data.disabled" (nzCheckedChange)="refreshStatus($event)"></td>
+          <td>编辑</td>
+          <td>{{data.name}}</td>
+          <td>{{data.buildingName}}</td>
+          <td>{{data.roomName}}</td>
+          <td>{{data.address}}</td>
+          <td>{{data.landArea}}</td>
+          <td>{{data.roomArea}}</td>
+          <td>{{data.price}}</td>
+          <td>{{data.isLoan}}</td>
+          <td>{{data.loanBank}}</td>
+          <td>{{data.agencyFee}}</td>
+          <td>{{data.isHavePark}}</td>
+          <td>{{data.progress}}</td>
+        </tr>
+      </tbody>
+    </nz-table>
+  </div>
+</div>
