+++ src/app/business/contract-manage/contract-manage.module.ts
@@ -1,10 +1,22 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ContractManageComponent } from './contract-manage.component';
+import { FormsModule } from '@angular/forms';
+import { ContractManageRoutesModule } from './contract-manage.routing';
+import {
+  NzTreeModule,
+  NzButtonModule,
+  NzTableModule
+} from 'ng-zorro-antd';
 
 @NgModule({
   imports: [
-    CommonModule
+    CommonModule,
+    FormsModule,
+    NzTreeModule,
+    NzButtonModule,
+    NzTableModule,
+    ContractManageRoutesModule,
   ],
   declarations: [ContractManageComponent]
 })
