+++ src/app/business/group-manage/group-manage.component.html
@@ -1,1 +1,39 @@
-楼盘管理模块
+<div class="group-manage-header">
+    <div class="function-module-name">
+        <span class="module-name">楼盘管理</span>
+    </div>
+    <div class="function-list">
+        <button nz-button nzType="primary">新增</button>
+        <button nz-button nzType="default">删除</button>
+    </div>
+    <div style="clear: both"></div>
+</div>
+<div class="group-manage-content">
+    <div class="search-result">共找到
+        <strong>{{groupTotal}}</strong>个楼盘</div>
+    <div class="search-content">
+        <nz-table class="na-table" #rowSelectionTable [nzData]="data" (nzCurrentPageDataChange)="currentPageDataChange($event)"
+            (nzPageIndexChange)="refreshStatus()" (nzPageSizeChange)="refreshStatus()">
+            <thead>
+                <tr>
+                    <th nzShowCheckbox [(nzChecked)]="allChecked" [nzIndeterminate]="indeterminate" (nzCheckedChange)="checkAll($event)"></th>
+                    <th>楼盘名称</th>
+                    <th>建筑物数量</th>
+                    <th>户室数量</th>
+                    <th>小区名</th>
+                    <th></th>
+                </tr>
+            </thead>
+            <tbody>
+                <tr *ngFor="let data of rowSelectionTable.data">
+                    <td nzShowCheckbox [(nzChecked)]="data.checked" [nzDisabled]="data.disabled" (nzCheckedChange)="refreshStatus($event)"></td>
+                    <td>{{data.name}}</td>
+                    <td>{{data.buildingCount}}</td>
+                    <td>{{data.roomCount}}</td>
+                    <td>{{data.courtName}}</td>
+                    <td>编辑</td>
+                </tr>
+            </tbody>
+        </nz-table>
+    </div>
+</div>
