+++ src/app/business/group-manage/group-manage.component.scss
@@ -1,0 +1,78 @@
+$group-manage-header-height:44px;
+$border-color:rgb(232, 232, 232);
+.host {
+    display: block;
+    width: 100%;
+    height: 100%;
+    overflow: hidden;
+    font-family: MicrosoftYaHei;
+}
+
+.group-manage-header {
+    width: 100%;
+    height: $group-manage-header-height;
+    border-bottom: 1px solid $border-color;
+    line-height: $group-manage-header-height;
+    font-family: MicrosoftYaHei;
+}
+
+.function-module-name,
+.group-name,
+.function-list {
+    float: left;
+    height: $group-manage-header-height;
+    padding: 0 10px;
+    line-height: $group-manage-header-height;
+}
+
+.function-module-name,
+.group-name {
+    color: #000;
+    font-weight: bold;
+}
+
+.function-module-name {
+    padding: 12px 0;
+    line-height: 20px;
+}
+
+.module-name {
+    display: block;
+    padding: 0 10px;
+    border-right: 1px solid #000;
+}
+
+button[nz-button] {
+    height: 28px;
+    margin-right: 8px;
+    line-height: 28px;
+}
+
+.group-manage-content {
+    float: left;
+    width: 100%;
+    height: calc(100% - #{$group-manage-header-height});
+    padding: 0 20px;
+}
+
+.search-result {
+    width: 100%;
+    height: 50px;
+    border-bottom: 1px solid $border-color;
+    line-height: 50px;
+    text-align: center;
+}
+
+.na-table {
+    ::ng-deep {
+        nz-pagination {
+            text-align: center;
+        }
+        .ant-pagination {
+            float: none;
+        }
+        th {
+            font-weight: bold;
+        }
+    }
+}
