--- src/app/business/realestate-admin/domains/contract-manage/contract-manage.module.ts
@@ -1,23 +1,0 @@
-import { FormsModule } from '@angular/forms';
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { ContractManageComponent } from './contract-manage.component';
-import { ContractManageRoutesModule } from './contract-manage.routing';
-import {
-  NzTreeModule,
-  NzButtonModule,
-  NzTableModule
-} from 'ng-zorro-antd';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    NzTreeModule,
-    NzButtonModule,
-    NzTableModule,
-    ContractManageRoutesModule,
-  ],
-  declarations: [ContractManageComponent]
-})
-export class ContractManageModule { }
